+++ src/main/java/org/opensearch/neuralsearch/query/HybridQueryBuilder.java
@@ -9,6 +9,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
@@ -23,6 +24,7 @@ import org.opensearch.core.common.io.stream.StreamOutput;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.index.query.AbstractQueryBuilder;
+import org.opensearch.index.query.InnerHitContextBuilder;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.index.query.QueryRewriteContext;
 import org.opensearch.index.query.QueryShardContext;
@@ -308,6 +310,20 @@ public final class HybridQueryBuilder extends AbstractQueryBuilder<HybridQueryBu
         QueryBuilderVisitor subVisitor = visitor.getChildVisitor(Occur.MUST);
         for (QueryBuilder subQueryBuilder : queries) {
             subQueryBuilder.visit(subVisitor);
+        }
+    }
+
+    /**
+     * Extracts the inner hits from the query tree
+     * @param innerHits
+     */
+    @Override
+    protected void extractInnerHitBuilders(Map<String, InnerHitContextBuilder> innerHits) {
+        if (Objects.isNull(queries)) {
+            return;
+        }
+        for (QueryBuilder query : queries) {
+            InnerHitContextBuilder.extractInnerHits(query, innerHits);
         }
     }
 }